+++ app/src/main/java/com/github/viktorpenelski/evtconf/fragments/TournamentFragment.kt
@@ -36,8 +36,6 @@ class TournamentFragment : Fragment() {
         initializeListViewOnItemClickListener(rootView)
         initializeSwipeToRefresh(rootView)
 
-        Log.d(TAG, "ON CREATE VIEW")
-
         return rootView
     }
 
@@ -53,7 +51,6 @@ class TournamentFragment : Fragment() {
                         loadExternalUrlWebView(entryClicked.redirectUrl)
                     }
 
-                    Log.d(TAG, "AdapterView<*>: $adapter \n View: $arg1 \n Int: $position \n Long: $arg3")
                 }
             }
         }
@@ -67,7 +64,6 @@ class TournamentFragment : Fragment() {
              * performs a swipe-to-refresh gesture.
              */
             setOnRefreshListener {
-                Log.i(TAG, "onRefresh called from SwipeRefreshLayout")
                 loadListView(view, true)
             }
         }