+++ app/src/main/java/com/github/viktorpenelski/evtconf/spreadsheet/tournaments/TournamentSpreadsheetEntryFactory.kt
@@ -27,7 +27,6 @@ class TournamentSpreadsheetEntryFactory : SpreadsheetEntryFactory<TournamentSpre
                     obj.getJSONObject(PICTURE).getString(VALUE),
                     obj.getJSONObject(LINK).getString(VALUE))
         } catch (e: JSONException) {
-            Log.e("deserialization", "Something went wrong during deserialization of remote json - ${e.stackTrace}")
             TournamentSpreadsheetEntry("", "", "")
         }
     }